Changed sat_1 directory to MapProjections and added in sat_2 code for finding high/low pressure values
